pyproject.toml
[project]
name = "station-nanna"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Thomas Bonnici", email = "thomas.bonnici@icloud.com"},
]
dependencies = ["fastapi>=0.115.8", "SQLAlchemy>=2.0.38", "uvicorn>=0.34.0", "python-vlc>=3.0.21203"]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pdm.scripts]
start = "uvicorn src.main:app --reload"

